pyproject.toml
[project]
name = "genai-archi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "cryptography==43.0.1",
    "distro==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httpx==0.27.2",
    "idna==3.8",
    "jiter==0.5.0",
    "openai==1.44.1",
    "pycparser==2.22",
    "pydantic==2.9.1",
    "pydantic-core==2.23.3",
    "python-dotenv==1.0.1",
    "sniffio==1.3.1",
    "tqdm==4.66.5",
    "typing-extensions==4.12.2",
]
